clutter-box-layout: Fix distribute_natural_allocation assertion
`size` is `gfloat` here. If it is `infinity`, the `MAX` macro evaluates
it to be greater than zero, but `distribute_natural_allocation` expects
a `gint`, so `size` becomes G_MININT.

This only fixes the assertion so the underlying bug of `size` being
`infinity` should still get fixed.

+++ clutter/clutter/clutter-box-layout.c
@@ -624,7 +624,7 @@ get_preferred_size_for_opposite_orientation (ClutterBoxLayout   *self,
   else
     {
       /* Bring children up to size first */
-      size = distribute_natural_allocation (MAX (0, size), nvis_children, sizes);
+      size = distribute_natural_allocation (MAX (0, (gint) size), nvis_children, sizes);
 
       /* Calculate space which hasn't distributed yet,
        * and is available for expanding children.
